+++ src/main/java/racingcar/dao/GameDao.java
@@ -1,30 +1,5 @@
 package racingcar.dao;
 
-import java.sql.PreparedStatement;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Component;
-
-@Component
-class GameDao {
-    private final JdbcTemplate jdbcTemplate;
-
-    public GameDao(final JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public int save(final int trialCount, final String winners) {
-        final String sql = "insert into game (trial, winners) values (?,?)";
-
-        final KeyHolder keyHolder = new GeneratedKeyHolder();
-        jdbcTemplate.update(connection -> {
-            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
-            ps.setInt(1, trialCount);
-            ps.setString(2, winners);
-            return ps;
-        }, keyHolder);
-
-        return keyHolder.getKey().intValue();
-    }
+public interface GameDao {
+    int save(final int trialCount, final String winners);
 }
