+++ src/main/java/racingcar/dao/GameJdbcDao.java
@@ -1,0 +1,30 @@
+package racingcar.dao;
+
+import java.sql.PreparedStatement;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Component;
+
+@Component
+public class GameJdbcDao implements GameDao {
+    private final JdbcTemplate jdbcTemplate;
+
+    public GameJdbcDao(final JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public int save(final int trialCount, final String winners) {
+        final String sql = "insert into game (trial, winners) values (?,?)";
+
+        final KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
+            ps.setInt(1, trialCount);
+            ps.setString(2, winners);
+            return ps;
+        }, keyHolder);
+
+        return keyHolder.getKey().intValue();
+    }
+}
